refactor(tauri): extract mcp and ipc helpers
