+++ src-tauri/src/lib.rs
@@ -8,14 +8,17 @@ mod error_window;
 mod idle;
 mod logging;
 mod mcp_server;
+mod mcp_token;
 mod model_management;
 mod monitor;
+mod monitor_ipc;
 mod native_messaging;
 mod power;
 mod python;
 mod registry_config;
 mod resource_utils;
 mod reverse_ipc;
+mod reverse_ipc_protocol;
 mod script_integrity;
 mod sensitive_filter;
 mod storage;
